requirements.txt
annotated-types==0.7.0
anyio==4.11.0
audioop-lts==0.2.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
comtypes==1.4.12
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.38.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
keyboard==0.13.5
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydantic==2.11.9
pydantic_core==2.33.2
pygame==2.6.1
pyparsing==3.2.5
pypiwin32==223
pyttsx3==2.99
pywin32==311
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
SpeechRecognition==3.14.3
standard-aifc==3.13.0
standard-chunk==3.13.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
webrtcvad-wheels==2.0.14
websockets==15.0.1
